% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.baseDotPlot}
\alias{.baseDotPlot}
\title{Dot Plot Base}
\usage{
.baseDotPlot(
  plotData,
  group = NULL,
  scale = NULL,
  AverageThreshold = NULL,
  NumDetectedThreshold = NULL,
  dotColors = NULL,
  fontSize = 11,
  fontFamily = "",
  flipPlot = FALSE
)
}
\arguments{
\item{group}{Column name from colData/metadata of the object to group cells by}

\item{scale}{Logical indicating whether the data should be scaled and centered}

\item{AverageThreshold}{Numeric specifying threshold for average expression,
where values below AverageThreshold and NumDetectedThreshold are transparent}

\item{NumDetectedThreshold}{Numeric specifying threshold for fraction of cells,
where values below AverageThreshold and NumDetectedThreshold are transparent}

\item{dotColors}{A character vector specifying the colors to be used
in the gradient for the dots. If scale is set to TRUE, the first 3 colors
will be used for the negative, zero, and positive values, respectively.}

\item{fontSize}{Numeric specifying the base font size in pts}

\item{fontFamily}{Character specifying the base font family}

\item{flipPlot}{Logical indicating whether to flip the x and y coordinates}
}
\value{
A ggplot2
}
\description{
Dot Plot Base
}
\keyword{internal}
