% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.createAnno}
\alias{.createAnno}
\title{Create column annotations}
\usage{
.createAnno(
  annoLabel,
  annoType = c("column", "row"),
  dotPlot = dotPlot,
  annoColors = NULL,
  groupLegends = TRUE,
  featureLegends = TRUE,
  fontSize = 11,
  fontFamily = "",
  flipPlot = FALSE
)
}
\arguments{
\item{annoColors}{A list with a name for each annotation that contains
a named vector of colors, where the name is the pairing of values to colors}

\item{groupLegends}{Logical indicating whether to show legends for group
annotations}

\item{featureLegends}{Logical indicating whether to show legends for feature
annotations}

\item{fontSize}{Numeric specifying the base font size in pts}

\item{fontFamily}{Character specifying the base font family}

\item{flipPlot}{Logical indicating whether to flip the x and y coordinates}
}
\value{
A ggplot2
}
\description{
Create column annotations
}
\keyword{internal}
