% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDotPlot-package.R
\docType{package}
\name{scDotPlot-package}
\alias{scDotPlot-package}
\title{scDotPlot: Cluster a Single-cell RNA-seq Dot Plot}
\description{
Dot plots of single-cell RNA-seq data allow for an examination of the relationships between cell groupings (e.g. clusters) and marker gene expression. The scDotPlot package offers a unified approach to perform a hierarchical clustering analysis and add annotations to the columns and/or rows of a scRNA-seq dot plot. It works with SingleCellExperiment and Seurat objects as well as data frames.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ben-laufer/scDotPlot}
  \item Report bugs at \url{https://github.com/ben-laufer/scDotPlot/issues}
}

}
\author{
\strong{Maintainer}: Benjamin I Laufer \email{blaufer@gmail.com}

Authors:
\itemize{
  \item Brad A Friedman
}

}
\keyword{internal}
